requirements.pip.txt
pyyaml
pytest
pytest-cov
fabric3